requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.7
python-multipart==0.0.6
PyPDF2==3.0.1
python-pptx==0.6.21
speechrecognition==3.10.0
pyaudio==0.2.11
websockets==11.0.3
asyncio==3.4.3
python-dotenv==1.0.0
werkzeug==2.3.7